refactor: generalize image-carousel files for npm
remove unnecessary style.css content and change a
class name for clarity.

modify index.js with a params section to describe
the parameters that the new wrapper function,
createImageCarousel(), accepts and help developers
understand how to use it, modify next/prevImage to
be more functionally complete, remove
eventListeners for the arrows, and add an API
section that gives simpler commands to execute
functions.

update README.md with complete instructions on how
to use the package.

build a new package.json file for npm publishing.